make TextureUsage and TargetBufferFlags enum class
This prevents values like NONE or COLOR which are very generic
names to collide with other classic enums. It also forces us to be
specific about what we're doing when converting to bitfields.

Also added a utility to easily enable any enum or enum class to
support binary operation (i.e. Bitmask contract), e.g.:

template<> struct utils::EnableBitMaskOperators<Foo> : public std::true_type{};

gives all binary operators to 'Foo'. Foo must be an enum.

Also added any() and none() convenience to convert an enum to a boolean.

+++ filament/backend/src/vulkan/VulkanDriver.cpp
@@ -709,8 +709,8 @@ void VulkanDriver::beginRenderPass(Handle<HwRenderTarget> rth,
         .colorFormat = color.format,
         .depthFormat = depth.format,
         .flags.clear         = params.flags.clear,
-        .flags.discardStart  = params.flags.discardStart,
-        .flags.discardEnd    = params.flags.discardEnd,
+        .flags.discardStart  = (uint8_t)params.flags.discardStart,
+        .flags.discardEnd    = (uint8_t)params.flags.discardEnd,
         .flags.dependencies  = params.flags.dependencies
     });
     mBinder.bindRenderPass(renderPass);
